Normalize page layouts
