+++ src/app/component/footer/_footer.scss
@@ -3,6 +3,7 @@
   text-align: center;
   border-top: 1px solid $color-border-minor;
   color: $color-minor;
+  background-color: $color-bg;
 
   &__grid {
     display: grid;
