+++ src/app/component/page-layout/page-layout.js
@@ -4,11 +4,22 @@ import React from 'react'
 import propTypes from '../../prop-types'
 import buildClassName from '../../lib/class-names'
 
-const PageLayout = ({classNames, children, header, footer, minorBackground = false}) => (
-  <div className={buildClassName('PageLayout', {minorBackground}, classNames)}>
-    <div className="PageLayout__header">{header}</div>
-    <main className="PageLayout__main">{children}</main>
-    {footer && <div className="PageLayout__footer">{footer}</div>}
+const PageLayout = ({
+  classNames,
+  children,
+  header,
+  stickyFooter,
+  footer,
+  showStickyFooter = false,
+  minorBackground = false
+}) => (
+  <div className={buildClassName('PageLayout', {showStickyFooter, minorBackground}, classNames)}>
+    <main className="PageLayout__main">
+      <div className="PageLayout__header">{header}</div>
+      <div className="PageLayout__content">{children}</div>
+      {stickyFooter && <div className="PageLayout__stickyFooter">{stickyFooter}</div>}
+    </main>
+    {footer}
   </div>
 )
 
@@ -17,6 +28,8 @@ PageLayout.propTypes = {
   children: PropTypes.node.isRequired,
   header: PropTypes.node.isRequired,
   footer: PropTypes.node,
+  stickyFooter: PropTypes.node,
+  showStickyFooter: PropTypes.bool,
   minorBackground: PropTypes.bool
 }
