+++ src/app/component/sticky-footer/sticky-footer.story.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import {storiesOf} from '@storybook/react'
+
+import StickyFooter from '.'
+
+storiesOf('StickyFooter', module).add('default', () => (
+  <StickyFooter>Sticky footer content</StickyFooter>
+))
