--- src/app/component/upload-layout/_upload-layout.scss
@@ -1,55 +1,0 @@
-.UploadLayout {
-  $stacking: content, sticky;
-
-  display: flex;
-  flex-direction: column;
-  position: relative;
-
-  &__header {
-    position: sticky;
-    top: 0;
-    flex-shrink: 0;
-    z-index: z($stacking, sticky);
-  }
-
-  &__main {
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
-    flex-shrink: 0;
-    flex-grow: 1;
-  }
-
-  &__content {
-    position: relative;
-    margin-top: $space-l;
-    flex-grow: 1;
-    height: 100%;
-    z-index: z($stacking, content);
-  }
-
-  &__stickyFooter {
-    display: none;
-    position: sticky;
-    bottom: 0;
-    background: $color-bg;
-    padding: $space-base $space-l;
-    text-align: right;
-    border-top: 1px solid $color-bg-minor;
-    z-index: z($stacking, sticky);
-
-    & > * + * {
-      margin-left: $space-xl;
-    }
-  }
-
-  &--hasModels {
-    .UploadLayout__stickyFooter {
-      display: block;
-    }
-
-    .UploadLayout__main {
-      background: $color-bg-minor;
-    }
-  }
-}
