--- src/app/component/upload-layout/index.js
@@ -1,1 +1,0 @@
-export {default} from './upload-layout'
